% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca.R
\docType{package}
\name{alpaca}
\alias{alpaca}
\alias{alpaca-package}
\title{alpaca: A package for fitting glm's with high-dimensional \eqn{k}-way fixed
effects.}
\description{
"ALternating Projection Algorithm for glm's with
multiple high-dimensional CAtegorical variables" (alpaca).

\strong{Keywords}: Generalized linear models, alternating projections,  and 
high-dimensional \eqn{k}-way fixed effects.
}
