% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsSameMetadata.R
\name{antsSameMetadata}
\alias{antsSameMetadata}
\title{Check if Two ANTs Images Have the Same Metadata}
\usage{
antsSameMetadata(ants1, ants2)
}
\arguments{
\item{ants1}{An ANTs image object.}

\item{ants2}{Another ANTs image object.}
}
\value{
A logical value indicating whether the two ANTs images have the same metadata. Returns TRUE if they match and FALSE otherwise.
}
\description{
This function checks if two ANTs images have the same metadata, including direction, origin, and spacing.
}
\examples{
\dontrun{
antsImage1 <- antsImageRead("image1.nii")
antsImage2 <- antsImageRead("image2.nii")
antsSameMetadata(antsImage1, antsImage2)
}

}
\seealso{
\code{\link{antsGetDirection}}, \code{\link{antsGetOrigin}}, \code{\link{antsGetSpacing}}
}
