% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.alpaca.R
\name{coef.alpaca}
\alias{coef.alpaca}
\title{Extract estimated Structural Parameters}
\usage{
\method{coef}{alpaca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{alpaca}.}

\item{...}{other arguments.}
}
\value{
The function \code{coef.alpaca} returns a named vector of estimated 
structural parameters.
}
\description{
\code{coef.alpaca} is a generic function which extracts
estimated structural parameters from objects returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
