% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.APEs}
\alias{coef.APEs}
\title{Extract estimates of average partial effects}
\usage{
\method{coef}{APEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"APEs"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{coef.APEs}} returns a named vector of estimates of the average partial 
effects.
}
\description{
\code{\link{coef.APEs}} is a generic function which extracts estimates of the average partial 
effects from objects returned by \code{\link{getAPEs}}.
}
\seealso{
\code{\link{getAPEs}}
}
