% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.feglm}
\alias{coef.feglm}
\title{Extract estimates of structural parameters}
\usage{
\method{coef}{feglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{feglm}.}

\item{...}{other arguments.}
}
\value{
The function \code{coef.feglm} returns a named vector of estimates refering to the structural
parameters.
}
\description{
\code{coef.feglm} is a generic function which extracts estimates of the structural parameters
from objects returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
