% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.summary.alpaca.R
\name{coef.summary.alpaca}
\alias{coef.summary.alpaca}
\title{Extract Matrix of estimated Structural Parameters}
\usage{
\method{coef}{summary.alpaca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{summary.alpaca}.}

\item{...}{other arguments.}
}
\value{
The function \code{coef.summary.alpaca} returns a named matrix of estimates 
related to the structural parameters.
}
\description{
\code{coef.summary.alpaca} is a generic function which extracts
the matrix of estimated structural parameters from objects returned by
\code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
