% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.summary.APEs}
\alias{coef.summary.APEs}
\title{Extract coefficient matrix of average partial effects}
\usage{
\method{coef}{summary.APEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"summary.APEs"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{coef.summary.APEs}} returns a named matrix of estimates related to the
average partial effects.
}
\description{
\code{\link{coef.summary.APEs}} is a generic function which extracts a coefficient matrix of 
average partial effects from objects returned by \code{\link{getAPEs}}.
}
\seealso{
\code{\link{getAPEs}}
}
