% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.summary.feglm}
\alias{coef.summary.feglm}
\title{Extract coefficient matrix of structural parameters}
\usage{
\method{coef}{summary.feglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{summary.feglm}.}

\item{...}{other arguments.}
}
\value{
The function \code{coef.summary.feglm} returns a named matrix of estimates related to the
structural parameters.
}
\description{
\code{coef.summary.feglm} is a generic function which extracts a coefficient matrix of structural
parameters from objects returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
