% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_patch.R
\name{extract_patch}
\alias{extract_patch}
\title{Extract Patch}
\usage{
extract_patch(
  candidate_id,
  patch_starts,
  patch_ends,
  t1,
  flair,
  epi,
  phase,
  labeled_candidates,
  rotate_patches
)
}
\arguments{
\item{candidate_id}{An identifier for the candidate lesion region.}

\item{patch_starts}{A vector specifying the starting coordinates of the patch in three dimensions.}

\item{patch_ends}{A vector specifying the ending coordinates of the patch in three dimensions.}

\item{t1}{antsImage representing the T1-weighted MRI image.}

\item{flair}{antsImage representing the FLAIR MRI image.}

\item{epi}{antsImage representing the EPI MRI image.}

\item{phase}{antsImage representing the phase MRI image.}

\item{labeled_candidates}{antsImage representing labeled candidates for lesion regions.}

\item{rotate_patches}{A logical flag indicating whether to rotate the extracted patches.}
}
\value{
A concatenated tensor containing the extracted and optionally rotated patches.
}
\description{
This function extracts a patch from various MRI images based on specified parameters,
optionally rotates the patches, and returns them as a concatenated tensor.
}
\examples{
\dontrun{
# Extract a patch with no rotation.
extracted_patch <- extract_patch(candidate_id = 1,
                                 patch_starts = c(10, 20, 30),
                                 patch_ends = c(20, 30, 40),
                                 t1 = t1_image,
                                 flair = flair_image,
                                 epi = epi_image,
                                 phase = phase_image,
                                 labeled_candidates = labeled_candidates_image,
                                 rotate_patches = FALSE)

# Extract a patch and apply random rotation.
extracted_patch <- extract_patch(candidate_id = 2,
                                 patch_starts = c(15, 25, 35),
                                 patch_ends = c(25, 35, 45),
                                 t1 = t1_image,
                                 flair = flair_image,
                                 epi = epi_image,
                                 phase = phase_image,
                                 labeled_candidates = labeled_candidates_image,
                                 rotate_patches = TRUE)
}
}
