% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm.R
\name{feglm}
\alias{feglm}
\title{feglm: A function to efficiently estimate glm's with high-dimensional 
\eqn{k}-way fixed effects}
\usage{
feglm(formula = NULL, data = NULL, beta.start = NULL,
  D.alpha.start = NULL, family = c("logit", "poisson"), control = list())
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of the
model to be fitted. \code{formula} must be of type \code{y ~ x | k},
where the \eqn{k} refers to an identifier of different levels in category
\eqn{k}.}

\item{data}{an object of class \code{"data.frame"} containing the variables in the
model.}

\item{beta.start}{an optional vector of starting values used for the structural
parameters in the linear predictor. Default see \code{Details}.}

\item{D.alpha.start}{an optional vector of starting values used for the fixed
effects contribution in the linear predictor. Default see \code{Details}.}

\item{family}{the description of the error distribution and link function to be used
in the model. This has to be a character string naming the
model function. The value should be any of \code{"logit"} or
\code{"poisson"}. Default is \code{"logit"}.}

\item{control}{a named list of parameters for controlling the fitting process.
Defaults see \code{\link{alpaca.control}}.}
}
\value{
The function \code{feglm} returns a named list of class \code{"alpaca"}.
}
\description{
\strong{Caution}: Package version 0.1 preliminary!

\code{feglm} is used to fit fixed effects generalized linear models with many
high-dimensional fixed effects based on unconditional maximum likelihood.
The algorithm can be interpreted as a pseudo demeaning approach that combines
the work of Gaure (2013) and Stammann et. al. (2016). For technical details
see Stammann (2018).

\strong{Remark:} The term fixed effect is used in econometrician's sense of
having intercepts for each level in each category.
}
\details{
...
}
\references{
Gaure, S. (2013). "OLS with Multiple High Dimensional Category Variables".
Computational Statistics and Data Analysis. 66.

Stammann, A., F. Heiss, and D. McFadden (2016). 
"Estimating Fixed Effects Logit Models with Large Panel Data". Working paper.

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear
Models with High-Dimensional K-Way Fixed Effects". Working Paper.
}
