% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.alpaca.R
\name{fitted.alpaca}
\alias{fitted.alpaca}
\title{Extract \code{feglm} Fitted Values}
\usage{
\method{fitted}{alpaca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{alpaca}.}

\item{...}{other arguments.}
}
\value{
The function \code{fitted.alpaca} returns a vector of fitted values.
}
\description{
\code{fitted.alpaca} is a generic function which extract fitted values from
an object returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
