% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.feglm}
\alias{fitted.feglm}
\title{Extract \code{feglm} fitted values}
\usage{
\method{fitted}{feglm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"feglm"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{fitted.feglm}} returns a vector of fitted values.
}
\description{
\code{\link{fitted.feglm}} is a generic function which extracts fitted values from an object 
returned by \code{\link{feglm}}.
}
\seealso{
\code{\link{feglm}}
}
