% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_model.R
\docType{data}
\name{mimosa_model}
\alias{mimosa_model}
\title{MIMoSA trained model for T1 and FLAIR, for ALPaCA use}
\usage{
mimosa_model
}
\description{
Trained MIMoSA model for when T1 and FLAIR images are available
}
\keyword{datasets,}
\keyword{model}
\keyword{trained,}
