% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.alpaca.R
\name{predict.alpaca}
\alias{predict.alpaca}
\title{Predict Method for \code{feglm} Fits}
\usage{
\method{predict}{alpaca}(object, type = c("link", "response"), ...)
}
\arguments{
\item{object}{an object of class \code{alpaca}.}

\item{type}{the type of prediction required. \code{"link"} is on the scale to the
linear predictor whereas \code{"response"} is on the scale of the response
variable. Default is \code{"link"}.}

\item{...}{other arguments.}
}
\value{
The function \code{predict.alpaca} returns a vector of predictions
}
\description{
\code{predict.alpaca} is a generic function which obtains predictions from
an object returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
