% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.feglm}
\alias{predict.feglm}
\title{Predict method for \code{feglm} fits}
\usage{
\method{predict}{feglm}(object, type = c("link", "response"), ...)
}
\arguments{
\item{object}{an object of class \code{"feglm"}.}

\item{type}{the type of prediction required. \code{"link"} is on the scale of the linear predictor
whereas \code{"response"} is on the scale of the response variable. Default is \code{"link"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{predict.feglm}} returns a vector of predictions.
}
\description{
\code{\link{predict.feglm}} is a generic function which obtains predictions from an object 
returned by \code{\link{feglm}}.
}
\seealso{
\code{\link{feglm}}
}
