% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.alpaca.R
\name{print.alpaca}
\alias{print.alpaca}
\title{Print \code{alpaca}}
\usage{
\method{print}{alpaca}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{alpaca}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{print.alpaca} is a generic function which displays some minimal
information from objects returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
