% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.feglm}
\alias{print.feglm}
\title{Print \code{feglm}}
\usage{
\method{print}{feglm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{feglm}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{print.feglm} is a generic function which displays some minimal information from objects
returned by \code{feglm}.
}
\seealso{
\code{\link{feglm}}
}
