% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.alpaca.R
\name{print.summary.alpaca}
\alias{print.summary.alpaca}
\title{Print \code{summary.alpaca}}
\usage{
\method{print}{summary.alpaca}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{an object of class \code{summary.alpaca}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{print.summary.alpaca} is a generic function which displays summary
statistics from objects returned by \code{summary.alpaca}.
}
\seealso{
\code{\link{feglm}}
}
