% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.APEs}
\alias{print.summary.APEs}
\title{Print \code{summary.APEs}}
\usage{
\method{print}{summary.APEs}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{an object of class \code{"summary.APEs"}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{\link{print.summary.APEs}} is a generic function which displays summary statistics from 
objects returned by \code{\link{summary.APEs}}.
}
\seealso{
\code{\link{getAPEs}}
}
