% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.feglm}
\alias{print.summary.feglm}
\title{Print \code{summary.feglm}}
\usage{
\method{print}{summary.feglm}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{an object of class \code{summary.feglm}.}

\item{digits}{unsigned integer indicating the number of decimal places. Default is 
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{other arguments.}
}
\description{
\code{print.summary.feglm} is a generic function which displays summary statistics from objects
returned by \code{summary.feglm}.
}
\seealso{
\code{\link{feglm}}
}
