% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_patch.R
\name{rotate_patch}
\alias{rotate_patch}
\title{Rotate Patch}
\usage{
rotate_patch(patch, invert, face, rotations)
}
\arguments{
\item{patch}{A tensor representing the patch to be rotated.}

\item{invert}{An integer flag indicating whether to invert the patch (1 for inversion, 0 for no inversion).}

\item{face}{An integer specifying the face of the cube to rotate towards (values from 1 to 6).}

\item{rotations}{An integer indicating the number of rotations (values from 0 to 3) to apply to the patch.}
}
\value{
A tensor representing the rotated patch.
}
\description{
This function rotates a patch based on specified parameters.
}
\examples{
\dontrun{
# Rotate the patch with no inversion, with the second face facedown, and performing one rotation.
rotated_patch <- rotate_patch(patch, invert = 0, face = 2, rotations = 1)

# Rotate the patch with inversion, with the third face facedown, and performing three rotations.
rotated_patch <- rotate_patch(patch, invert = 1, face = 3, rotations = 3)
}
}
