% Generated by roxyassert: do not edit by hand.
\name{roxyassert-generated-asserts}
\alias{roxyassert-generated-asserts}
\alias{assert_type_Account}
\alias{assert_type_AccountConfig}
\alias{assert_type_Activity}
\alias{assert_type_Asset}
\alias{assert_type_Bars}
\alias{assert_type_BarsMulti}
\alias{assert_type_Calendar}
\alias{assert_type_Clock}
\alias{assert_type_Contract}
\alias{assert_type_CorporateAction}
\alias{assert_type_CryptoOrderbook}
\alias{assert_type_MostActives}
\alias{assert_type_Movers}
\alias{assert_type_News}
\alias{assert_type_OptionQuotesMulti}
\alias{assert_type_OptionTradesMulti}
\alias{assert_type_Order}
\alias{assert_type_OrderCore}
\alias{assert_type_PortfolioHistory}
\alias{assert_type_Position}
\alias{assert_type_Quote}
\alias{assert_type_QuotesMulti}
\alias{assert_type_Snapshot}
\alias{assert_type_SnapshotMulti}
\alias{assert_type_Trade}
\alias{assert_type_TradesMulti}
\alias{assert_type_Watchlist}
\alias{assert_type_Watchlists}
\title{Generated assertion helpers}
\usage{
assert_type_Account(value)

assert_type_AccountConfig(value)

assert_type_Activity(value)

assert_type_Asset(value)

assert_type_Bars(value)

assert_type_BarsMulti(value)

assert_type_Calendar(value)

assert_type_Clock(value)

assert_type_Contract(value)

assert_type_CorporateAction(value)

assert_type_CryptoOrderbook(value)

assert_type_MostActives(value)

assert_type_Movers(value)

assert_type_News(value)

assert_type_OptionQuotesMulti(value)

assert_type_OptionTradesMulti(value)

assert_type_Order(value)

assert_type_OrderCore(value)

assert_type_PortfolioHistory(value)

assert_type_Position(value)

assert_type_Quote(value)

assert_type_QuotesMulti(value)

assert_type_Snapshot(value)

assert_type_SnapshotMulti(value)

assert_type_Trade(value)

assert_type_TradesMulti(value)

assert_type_Watchlist(value)

assert_type_Watchlists(value)
}
\arguments{
\item{value}{The value or shaped object validated by the assert helper.}
}
\value{
\code{assert_args_*} return \code{invisible(NULL)}; \code{assert_return_*}
return the validated value; \code{assert_type_*} return it invisibly. All are
called for the side effect of signalling an error on a contract violation.
}
\description{
Runtime assertion helpers generated by roxyassert from typed roxygen
documentation and exported via the \verb{@exportassert} tag.
}
\keyword{internal}
