% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Generate a Simulated Data Set}
\usage{
simData(n, t, seed)
}
\arguments{
\item{n}{number of cross-sectional units.}

\item{t}{number of time periods.}

\item{seed}{a seed to ensure reproducibility.}
}
\value{
The function \code{simData} returns a data.frame with 6 variables.
}
\description{
Constructs a simulated data set with \eqn{n} cross-sectional units observed
for \eqn{t} time periods according to the following data generating process:

\deqn{y_{it} = \mathbf{1}[\mathbf{x}_{it}^{\prime} \boldsymbol{\beta} +
 \alpha_{i} + \gamma_{t} + \epsilon_{it} > 0]\, ,}{y = 1[X \beta + \alpha + 
 \gamma + \epsilon],}
where \eqn{\mathbf{X}}{X} consists of 3 independent standard normally
distributed regressors and \eqn{\epsilon_{it}}{\epsilon} is an iid. logistic
error term with location zero and scale one. \eqn{\alpha_{i}}{\alpha} and 
\eqn{\gamma_{t}}{\gamma} are generated as iid. standard normal and 
\eqn{\boldsymbol{\beta} = [1, - 1, 1]^{\prime}}{\beta = [1, - 1, 1]'}.
}
