% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_custom_lesion_mask.r
\name{split_custom_lesion_mask}
\alias{split_custom_lesion_mask}
\title{Perform lesion splitting on custom lesion probability map}
\usage{
split_custom_lesion_mask(
  custom_map_path,
  custom_map_threshold,
  output_dir,
  verbose = FALSE,
  return_images = TRUE
)
}
\arguments{
\item{custom_map_path}{Path to the custom probability map path (in .nii.gz format).}

\item{custom_map_threshold}{Value from 0 to 1 indicating threshold for voxel segmentation.}

\item{output_dir}{Directory where preprocessed images and results will be saved.}

\item{verbose}{Logical, indicating whether to display verbose output (default is FALSE).}

\item{return_images}{Logical, indicating whether a named list of output images should be returned in addition to writing to disk (default is TRUE)}
}
\value{
Saves the following images to disk: labeled_candidates.nii.gz, eroded_candidates.nii.gz
If return_images = TRUE, also returns named list containing the images with names: labeled_candidates.nii.gz, eroded_candidates.nii.gz.
}
\description{
This function allows for lesion splitting of a custom lesion probability
map. The probability map must be continuous from 0 to 1 and is originally
designed to work with MIMoSA probability maps. If deep learning
probability maps are used instead, range-constrained non-sigmoid
transformed outputs may work better, as sigmoid transforms may place
probability values too close to 0 and 1 for the Hessian-based approach
to appropriately split. The lesion splitting algorithm is from "An
Automated Statistical Technique for Counting Distinct Multiple Sclerosis
Lesions" (2018) by Jordan Dworkin et al.
}
\examples{
\dontrun{
preprocess_images(
  "t1_image.nii.gz", "flair_image.nii.gz",
  "epi_image.nii.gz", "phase_image.nii.gz",
  "output_directory"
)
}
}
