% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.alpaca.R
\name{summary.alpaca}
\alias{summary.alpaca}
\title{Summarizing Models of Class \code{alpaca}}
\usage{
\method{summary}{alpaca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{alpaca}.}

\item{...}{other arguments.}
}
\value{
Returns an object of class \code{summary.alpaca} which is a list of summary
statistics of \code{object}.
}
\description{
Summary statistics for objects of class \code{alpaca}.
}
\seealso{
\code{\link{feglm}}
}
