% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.APEs}
\alias{summary.APEs}
\title{Summarizing models of class \code{APEs}}
\usage{
\method{summary}{APEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"APEs"}.}

\item{...}{other arguments.}
}
\value{
Returns an object of class \code{"summary.APEs"} which is a list of summary statistics of 
\code{object}.
}
\description{
Summary statistics for objects of class \code{"APEs"}.
}
\seealso{
\code{\link{getAPEs}}
}
