% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.feglm}
\alias{tidy.feglm}
\alias{feglm_tidiers}
\alias{alpaca_tidiers}
\title{Tidy a(n) feglm object}
\usage{
\method{tidy}{feglm}(
  x,
  fe = FALSE,
  type = "hessian",
  cluster = NULL,
  conf.int = NULL,
  conf.level = NULL,
  ...
)
}
\arguments{
\item{x}{A `feglm` object returned from [alpaca::feglm()].}

\item{fe}{Logical indicating whether or not to include
estimates of fixed effects. Defaults to `FALSE`.}

\item{type}{Character indicating the type of covariance
estimate to be used for the standard errors. Possible values
are `hessian`, `outer.product`, `sandwich`, `cluster`. See
alpaca::summary.feglm for details. Defaults to `hessian`.}

\item{cluster}{a symbolic description indicating the
clustering of observations.}

\item{conf.int}{Logical indicating whether or not to include a
confidence interval in the tidied output. Defaults to
`FALSE`. NOT IMPLEMENTED YET.}

\item{conf.level}{The confidence level to use for the
confidence interval if `conf.int = TRUE`. Must be strictly
greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval. NOT
IMPLEMENTED YET.}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. **Cautionary note:** Misspelled arguments will be
absorbed in `...`, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass `conf.lvel = 0.9`, all computation will
proceed using `conf.level = 0.95`. Additionally, if you pass
`newdata = my_tibble` to an [augment()] method that does not
accept a `newdata` argument, it will use the default value for
the `data` argument.}
}
\value{
A [tibble::tibble()] with one row for each term in the
  regression. The tibble has columns:

  \item{term}{The name of the regression term.}
  \item{estimate}{The estimated value of the regression term.}
  \item{std.error}{The standard error of the regression term.}
  \item{statistic}{The value of a statistic, almost always a
    T-statistic, to use in a hypothesis that the regression
    term is non-zero.}
  \item{p.value}{The two-sided p-value associated with the
    observed statistic.}
  \item{conf.low}{The low end of a confidence interval for the
    regression term. Included only if `conf.int = TRUE`.}
  \item{conf.high}{The high end of a confidence interval for
    the regression term. Included only if `conf.int = TRUE`.}
}
\description{
Tidy summarizes information about the components
  of a model. A model component might be a single term in a
  regression, a single hypothesis, a cluster, or a class.
  Exactly what tidy considers to be a model component varies
  across models but is usually self-evident. If a model has
  several distinct types of components, you will need to
  specify which components to return.
}
\examples{

library(alpaca)

dt <- simGLM(n = 40, t = 10, seed = 123, model = "poisson")

res_feglm <- feglm(formula = y ~ x1 + x2 + x3 | i + t,
                   data = dt, family = poisson())
res_feglm
tidy(res_feglm)
tidy(res_feglm, type = "sandwich")
tidy(res_feglm, type = "clustered", cluster = ~ i + t)
augment(res_feglm)
glance(res_feglm)

}
\seealso{
[alpaca::feglm()]

Other feglm tidiers: 
\code{\link{augment.feglm}()}
}
\concept{feglm tidiers}
