% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.alpaca.R
\name{vcov.alpaca}
\alias{vcov.alpaca}
\title{Extract estimated Covariance Matrix of the estimated Structural Parameters}
\usage{
\method{vcov}{alpaca}(object, ...)
}
\arguments{
\item{object}{an object of class \code{alpaca}.}

\item{...}{other arguments.}
}
\value{
The function \code{vcov.alpaca} returns a named matrix.
}
\description{
\code{vcov.alpaca} extracts the estimated covariance matrix of the estimated
structural parameters from objects returned by \code{feglm}. The estimation
is equal to inverse of Hessian after convergence.
}
\seealso{
\code{\link{feglm}}
}
