% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov.APEs}
\alias{vcov.APEs}
\title{Compute covariance matrix after estimating \code{APEs}}
\usage{
\method{vcov}{APEs}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"APEs"}.}

\item{...}{other arguments.}
}
\value{
The function \code{\link{vcov.APEs}} returns a named matrix of covariance estimates.
}
\description{
\code{\link{vcov.APEs}} estimates the covariance matrix for the estimator of the
average partial effects from objects returned by \code{\link{getAPEs}}.
}
\seealso{
\code{\link{getAPEs}}
}
