% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorr.R
\name{biasCorr}
\alias{biasCorr}
\title{Asymptotic bias correction after fitting binary choice models with a two-/three-way error
component}
\usage{
biasCorr(object = NULL, L = 0L, panel.structure = c("classic", "network"))
}
\arguments{
\item{object}{an object of class \code{"feglm"}; currently restricted to \code{\link[stats]{binomial}} 
with \code{"logit"} or \code{"probit"} link function.}

\item{L}{unsigned integer indicating a bandwidth for the estimation of spectral densities proposed by 
Hahn and Kuersteiner (2011). Default is zero, which should be used if all regressors are 
assumed to be strictly exogenous with respect to the error term. In the presence of weakly exogenous 
or predetermined regressors, Fernández-Val and Weidner (2016, 2018) suggest to choose a bandwidth 
zero and four. Note that the order of factors to be partialed out is important for bandwidths larger 
than zero (see vignette for details).}

\item{panel.structure}{a string equal to \code{"classic"} or \code{"network"} which determines the 
structure of the panel used. \code{"classic"} denotes panel structures where for example the same
cross-sectional units are observed several times (this includes pseudo panels). 
\code{"network"} denotes panel structures where for example bilateral trade flows are observed 
for several time periods. Default is \code{"classic"}.}
}
\value{
The function \code{\link{biasCorr}} returns a named list of classes \code{"biasCorr"} and 
\code{"feglm"}.
}
\description{
\code{\link{biasCorr}} is a post-estimation routine that can be used to substantially reduce the 
incidental parameter bias problem (Neyman and Scott (1948)) present in non-linear fixed effects 
models (see Fernández-Val and Weidner (2018) for an overview). The command applies the analytical 
bias correction derived by Fernández-Val and Weinder (2016) and Hinz, Stammann, and Wanner (2020) 
to obtain bias-corrected estimates of the structural parameters and is currently restricted to 
logit and probit models with two- and three-way fixed effects.
}
\examples{
\donttest{
# Generate an artificial data set for logit models
library(alpaca)
data <- simGLM(1000L, 20L, 1805L, model = "logit")

# Fit 'feglm()'
mod <- feglm(y ~ x1 + x2 + x3 | i + t, data)

# Apply analytical bias correction
mod.bc <- biasCorr(mod)
summary(mod.bc)
}
}
\references{
Czarnowske, D. and A. Stammann (2020). "Fixed Effects Binary Choice Models: Estimation and Inference
with Long Panels". ArXiv e-prints.

Fernández-Val, I. and M. Weidner (2016). "Individual and time effects in nonlinear panel models 
with large N, T". Journal of Econometrics, 192(1), 291-312.

Fernández-Val, I. and M. Weidner (2018). "Fixed effects estimation of large-t panel data 
models". Annual Review of Economics, 10, 109-138.

Hahn, J. and G. Kuersteiner (2011). "Bias reduction for dynamic nonlinear panel models with 
fixed effects". Econometric Theory, 27(6), 1152-1191.

Hinz, J., A. Stammann, and J. Wanner (2020). "State Dependence and Unobserved Heterogeneity
in the Extensive Margin of Trade". ArXiv e-prints.

Neyman, J. and E. L. Scott (1948). "Consistent estimates based on partially consistent 
observations". Econometrica, 16(1), 1-32.
}
\seealso{
\code{\link{feglm}}
}
