% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{#' Make Predictions using ALPaCA}
\usage{
make_predictions(
  ants_list = NULL,
  t1 = NULL,
  flair = NULL,
  epi = NULL,
  phase = NULL,
  labeled_candidates = NULL,
  lesion_priority = c("Youden's J", "Specificity", "Sensitivity"),
  prl_priority = c("Youden's J", "Specificity", "Sensitivity"),
  cvs_priority = c("Youden's J", "Specificity", "Sensitivity"),
  return_raw_probabilities = FALSE,
  clear_discordant_predictions = TRUE,
  n_patches = 20,
  n_models = 10,
  rotate_patches = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{ants_list}{An optional named list containing input images or file paths to images. Recommended to be output from the preprocess_images() function. Names must be: t1, flair, epi, phase, and labeled_candidates. Either ants_list must be provided, or t1, flair, epi, phase, and labeled_candidates must be provided.}

\item{t1}{antsImage or file path to .nii.gz representing T1-weighted MRI image.}

\item{flair}{antsImage or file path to .nii.gz representing the FLAIR MRI image.}

\item{epi}{antsImage or file path to .nii.gz representing the EPI MRI image.}

\item{phase}{antsImage or file path to .nii.gz representing the phase MRI image.}

\item{labeled_candidates}{antsImage or file path to .nii.gz representing labeled candidates for lesion regions.}

\item{lesion_priority}{A character vector specifying priority for lesion prediction thresholds -- Youden's J, Specificity, Sensitivity. Thresholds are based on training set ROC curves from CV models. Default priority is Youden's J, with sensitivity \eqn{\approx} 0.83 and specificity \eqn{\approx} 0.86. 'Specificity' prioritizes specificity 3 times more than sensitivity, with sensitivity \eqn{\approx} 0.69 and specificity \eqn{\approx} 0.94. 'Sensitivity' prioritizes sensitivity 3 times more than specificity, with sensitivity \eqn{\approx} 0.92 and specificity \eqn{\approx} 0.70.}

\item{prl_priority}{A character vector specifying priority for PRL prediction thresholds. Same options and default as lesion_priority. For Youden's J, sensitivity \eqn{\approx} 0.76 and specificity \eqn{\approx} 0.83. For Specificity, sensitivity \eqn{\approx} 0.63 and specificity \eqn{\approx} 0.90. For Sensitivity, sensitivity \eqn{\approx} 0.86 and specificity \eqn{\approx} 0.64.}

\item{cvs_priority}{A character vector specifying priority for CVS prediction thresholds.Same options and default as lesion_priority. For Youden's J, sensitivity \eqn{\approx} 0.81 and specificity \eqn{\approx} 0.65. For Specificity, sensitivity \eqn{\approx} 0.27 and specificity \eqn{\approx} 0.91. For Sensitivity, sensitivity \eqn{\approx} 0.91 and specificity \eqn{\approx} 0.47.}

\item{return_raw_probabilities}{A logical flag indicating whether to return raw probability antsImages for each region. A raw probability lesion-wise dataframe is always returned. (Default is FALSE)}

\item{clear_discordant_predictions}{A logical flag indicating whether to clear discordant predictions (ie candidates where the model predicts "CVS"/"PRL, but not "lesion".) In training, lesion prediction was almost always more reliable under Youden's J threshold choice. (Default is TRUE)}

\item{n_patches}{An integer specifying the number of patches to sample for predictions. Coordinates are sampled from within each lesion and a patch is built around that center coordinate. (Default is 20)}

\item{n_models}{An integer specifying the number of cross-validation models to use for predictions (between 1 and 10). Default is to use all CV models to predict and average the final prediction across all 10 models.}

\item{rotate_patches}{A logical flag indicating whether to rotate the patches used for predictions. Useful for decreasing dependence of predictions for each sample. (Default is TRUE)}

\item{verbose}{A logical flag indicating whether to display verbose progress messages. (Default is FALSE)}
}
\value{
A list containing the ALPaCA mask (segmentation of lesions), predictions, and prediction uncertainties.
}
\description{
This function runs the pre-trained Automated Lesion, PRL (Paramagnetic Rim Lesion), and CVS (Central Vein Sign) Analysis network. This model takes in multi-modal data in the form of pre-processed T1, FLAIR, EPI, and phase images to generate predictions of whether lesion candidates (identified via MIMoSA) are true lesions, PRLs, or CVS.
}
\examples{
\dontrun{
# Make predictions using input images and default parameters.
predictions <- make_predictions(t1 = t1_image,
                                 flair = flair_image,
                                 epi = epi_image,
                                 phase = phase_image,
                                 labeled_candidates = labeled_candidates_image)

# Make predictions using input images and return raw probabilities.
predictions <- make_predictions(t1 = t1_image,
                                 flair = flair_image,
                                 epi = epi_image,
                                 phase = phase_image,
                                 labeled_candidates = labeled_candidates_image,
                                 return_raw_probabilities = TRUE)
}
}
